renames some files
